requirements.txt
requests>=2.28.0
rich>=13.0.0
click>=8.0.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
phonenumbers>=8.13.0
python-whois>=0.9.3
dnspython>=2.3.0
holehe>=1.4.0
trio>=0.22.0
httpx>=0.28.0
python-dotenv>=1.0.0
